Re-attach resized disk with SSD TRIM feature enabled.
This allows to shrink the .vdi file on the host when disk space
is reclaimed in the guest OS.

For some reason VirtualBox doesn't expose this feature on its
GUI.  It's only available in VBoxManage when attaching a
virtual hard disk.  So this plugin is the perfect place to
enable this feature.

More information about this feature here:
https://www.virtualbox.org/manual/ch08.html

Look at the `storageattach --discard` option.

I've been using it for many months on my Ubuntu guest VMs with
no ill side-effects.